HSD #1508104447-11: mtd: nand: raw: denali: Fix comparison of distinct pointer types warning
Fix compilation warning below:

In file included from include/linux/compat.h:7:0,
                 from include/dm/device.h:17,
                 from include/dm.h:9,
                 from drivers/mtd/nand/raw/denali.c:8:
drivers/mtd/nand/raw/denali.c: In function ‘nand_spl_load_image’:
include/linux/kernel.h:177:17: warning: comparison of distinct pointer types lacks a cast
  (void) (&_min1 == &_min2);  \
                 ^
drivers/mtd/nand/raw/denali.c:1414:11: note: in expansion of macro ‘min’
   count = min(len, count);
           ^~~

+++ drivers/mtd/nand/raw/denali.c
@@ -1410,7 +1410,7 @@ int nand_spl_load_image(u32 offset, u32 len, void *dst)
 			return err;
 
 		count -= page_align_overhead;
-		count = min(len, count);
+		count = min((size_t)len, (size_t)count);
 		memcpy(dst, page_buffer + page_align_overhead, count);
 		free(page_buffer);
